backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Denys Shevchenko", email = "denshevchenko19@gmail.com" },
    { name = "Yehor Karabanov", email = "yehorkarabanov@gmail.com" }
]

requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "sqlalchemy[asyncio] (>=2.0.37,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "websockets (>=14.2,<15.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "inflect (>=7.5.0,<8.0.0)",
    "bcrypt (>=4.2.1,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
